fix type casting warnings between types of different sizes

+++ sscanApp/src/saveData.c
@@ -1234,7 +1234,7 @@ LOCAL int monitorScan(SCAN* pscan, int pass)
     for (i=0; i<SCAN_NBP; i++) {
         if (pscan->cpxnv[i]!=NULL && pscan->cpxsm[i]!=NULL && pscan->crxnv[i]!=NULL) {
             if (ca_add_event(DBR_SHORT, pscan->cpxnv[i], pxnvMonitor,
-                    (void*)(long)i, 0)!=ECA_NORMAL) {
+                    (void*)(size_t)i, 0)!=ECA_NORMAL) {
                 printf("Unable to monitor %s\n", ca_name(pscan->cpxnv[i]));
                 return -1;
             }
@@ -1244,7 +1244,7 @@ LOCAL int monitorScan(SCAN* pscan, int pass)
                 return -1;
             }
             if (ca_add_event(DBR_SHORT, pscan->crxnv[i], rxnvMonitor,
-                    (void*)(long)i, 0)!=ECA_NORMAL) {
+                    (void*)(size_t)i, 0)!=ECA_NORMAL) {
                 printf("Unable to monitor %s\n", ca_name(pscan->crxnv[i]));
                 return -1;
             }
@@ -1257,7 +1257,7 @@ LOCAL int monitorScan(SCAN* pscan, int pass)
     for (i=0; i<SCAN_NBD; i++) {
         if (pscan->cdxnv[i]!=NULL) {
             if (ca_add_event(DBR_SHORT, pscan->cdxnv[i], dxnvMonitor,
-                    (void*)(long)i, 0)!=ECA_NORMAL) {
+                    (void*)(size_t)i, 0)!=ECA_NORMAL) {
                 printf("Unable to monitor %s\n", ca_name(pscan->cdxnv[i]));
                 return -1;
             }
@@ -1270,12 +1270,12 @@ LOCAL int monitorScan(SCAN* pscan, int pass)
     for (i=0; i<SCAN_NBT; i++) {
         if (pscan->ctxnv[i]!=NULL && pscan->ctxcd[i]!=NULL) {
             if (ca_add_event(DBR_SHORT, pscan->ctxnv[i], txnvMonitor,
-                    (void*)(long)i, 0)!=ECA_NORMAL) {
+                    (void*)(size_t)i, 0)!=ECA_NORMAL) {
                 printf("Unable to monitor %s\n", ca_name(pscan->ctxnv[i]));
                 return -1;
             }
             if (ca_add_event(DBR_FLOAT, pscan->ctxcd[i], txcdMonitor,
-                    (void*)(long)i, 0)!=ECA_NORMAL) {
+                    (void*)(size_t)i, 0)!=ECA_NORMAL) {
                 printf("Unable to monitor %s\n", ca_name(pscan->ctxcd[i]));
                 return -1;
             }
@@ -1909,8 +1909,8 @@ LOCAL int connectPV(char* pv, char* desc)
     char buff[PVNAME_STRINGSZ];
     long  count;
     int   type;
-    int   len;
-    long  size;
+    size_t len;
+    size_t size;
 
     /* allocate space for the new pv */
     pnode= (PV_NODE*) malloc(sizeof(PV_NODE));
